requirements.txt
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
filelock==3.25.2
Flask==3.1.3
fsspec==2026.3.0
google-auth==2.49.2
google-genai==1.72.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mongoengine==0.29.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
pillow==12.1.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.17.0
python-dotenv==1.2.2
requests==2.33.1
setuptools==81.0.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.4
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.8
